% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEPlot.R
\name{CEPlot}
\alias{CEPlot}
\title{Cumulative eigenvalue plot}
\usage{
CEPlot(x, nDraw = NULL)
}
\arguments{
\item{x}{...}

\item{nDraw}{...}
}
\description{
Compares eigenvalues by cumulative Bland-Altman type plots.
Kolmogorov-Smirnov and Cramer-von Mises tests evaluated by permutation tests
}
\examples{
#load iris data
data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
iris$group <- "setosa"
iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
irisPCADSC <- PCADSC(iris, "group", setdiff(names(iris), c("group", "Species")))

#make a CE plot
CEPlot(irisPCADSC)

}

