% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumeigen.R
\name{cumeigen}
\alias{cumeigen}
\title{Cumulative eigenvalue plot}
\usage{
cumeigen(data, splitBy, var = NULL, B = 1000, make.plot = TRUE)
}
\arguments{
\item{data}{A dataset, either a \code{data.frame} or a \code{matrix} with variables
in columns and observations in rows. WHAT TO DO ABOUT DATA.TABLES AND ALSO TIBBLES. BOTH WILL BE
PROBLEMATIC AS OF NOW.}

\item{splitBy}{A grouping variable with two levels defining the two groups within the
dataset whose data structures we wish to compare. If \code{splitBy} has more than
two levels, only the first two levels are used.}

\item{var}{The variable names in \code{data} to include in the PCADSC. If \code{NULL}
(the default), all variables except for \code{splitBy} are used.}

\item{B}{Number of simulations}

\item{make.plot}{Should a plot be generated?}
}
\description{
Compares eigenvalues by cumulative Bland-Altman type plots. Kolmogorov-Smirnov and Cramer-von Mises tests evaluated by permutation tests
}
\examples{
#load iris data
data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
iris$group <- "setosa"
iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
cumeigen(iris, "group",var=setdiff(names(iris), c("group", "Species")))

}

