% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadComp.R
\name{loadComp}
\alias{loadComp}
\title{importFrom stats na.omit princomp
importFrom reshape2 melt}
\usage{
loadComp(data, vars, varCO = NULL)
}
\description{
importFrom stats na.omit princomp
importFrom reshape2 melt
}

