% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADSC.R
\docType{class}
\name{PCADSC-class}
\alias{PCADSC-class}
\title{A PCADSC object}
\description{
Blabla
}
\examples{
#load iris data
  data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
 iris$group <- "setosa"
 iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
 irisPCADSC <- makePCADSC(iris, "group",
                        var=setdiff(names(iris), c("group", "Species")))
#plot it
 qplot(irisPCADSC)

#print it
 print(irisPCADSC)

}

