% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADSC.R
\name{PCADSC}
\alias{PCADSC}
\title{Compute the elements used for PCADSC}
\usage{
PCADSC(data, splitBy, vars = NULL, doCE = TRUE, doAngle = TRUE,
  doChroma = TRUE, B = 1000)
}
\arguments{
\item{data}{A dataset, either a \code{data.frame} or a \code{matrix} with variables
in columns and observations in rows. WHAT TO DO ABOUT DATA.TABLES AND ALSO TIBBLES. BOTH WILL BE
PROBLEMATIC AS OF NOW.}

\item{splitBy}{A grouping variable with two levels defining the two groups within the
dataset whose data structures we wish to compare. If \code{splitBy} has more than
two levels, only the first two levels are used.}

\item{vars}{The variable names in \code{data} to include in the PCADSC. If \code{NULL}
(the default), all variables except for \code{splitBy} are used.}

\item{doCE}{...}

\item{doAngle}{...}

\item{doChroma}{...}

\item{B}{...}
}
\value{
A list including the following objects: ...
}
\description{
Calculate standardized PCA loadings and accumulated variance contributions
separately on two parts of a dataset in order to perform non-parametric data structure
comparisons.
}
\details{
Blablabla.
Note that \code{\link[tibble]{tibble}}s and \code{\link[data.table]{data.table}}s are
accepted as input, but they are instantly converted to \code{\link{data.frame}}s. Future
releases might include specific implementation for these data representations.
}
\examples{

#load iris data
data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
iris$group <- "setosa"
iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
irisPCADSC <- PCADSC(iris, "group", setdiff(names(iris), c("group", "Species")))


}
\seealso{
\code{\link{doCE}}, \code{\link{doAngle}}, \code{\link{doChroma}}
}

