% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.R
\name{plotPCADSC}
\alias{plotPCADSC}
\title{Plot a PCADSC object}
\usage{
plotPCADSC(x, varLabels = NULL, covCO = NULL, splitLabels = NULL)
}
\arguments{
\item{varLabels}{A vector of character string labels for the variables used in
\code{pcadscObj}. If non-null, these labels appear in the plot instead of the
variable names. Note that they must be listed in the same order as the variables
in \code{pcadscObj} and this order can be expected by calling
\code{pcadscObj$varNames}. If \code{NULL} (the default), these variable
names are used.}

\item{covCO}{Not implemented, something like a variance contribution cut-off, maybe?}

\item{splitLabels}{Labels for the two categories of the splitting variable used
to create the PCADSC object, \code{pcadscObj}, given as a named list (see examples).
These labels will appear in the headers of the two PCADSC plots.
If \code{NULL} (the default), the original levels of the splitting variable
are used.}

\item{pcadscObj}{A PCADSC object (see \code{\link{PCADSC}}).}
}
\value{
A PCADSC plot.
}
\description{
Given af PCADSC object, \code{pcadscObj}, (as produced by \code{\link{PCADSC}}), a data
structure comparison plot is generated in the style of  \code{ggplot2}.
}
\examples{
#Example showing how to use splitLabels

}
\seealso{
\code{\link{PCADSC}}
}

