% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printPCADSC}
\alias{printPCADSC}
\title{Print a PCADSC object}
\usage{
printPCADSC(x, ...)
}
\arguments{
\item{x}{A PCADSC object}

\item{...}{For internal use only}
}
\description{
Print an overview of the results of PCADSC, as represented by a PCADSC object generated
by \code{\link{makePCADSC}}.
}
\examples{

#load iris data
data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
iris$group <- "setosa"
iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
irisPCADSC <- makePCADSC(iris, "group",
                        var=setdiff(names(iris), c("group", "Species")))
print(irisPCADSC)

}
\seealso{
\code{\link{makePCADSC}} \code{\link{plotPCADSC}}
}

