% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQanglePlot.R
\name{QQanglePlot}
\alias{QQanglePlot}
\title{QQ-plot for anglePlot}
\usage{
QQanglePlot(x)
}
\arguments{
\item{x}{A \code{PCADSC} object with angle information simulated under the null.}
}
\description{
Quantile-quantile plot for the observed anglePlot p-values against
the null hypothesis of equal structure. Size of points display the length of the
arrows in the anglePlot. Also shown is pointwise 95pct confidence interval under
the null together with 20 random samples under the null. Finally, the p-value for
one-sided Kolmogorov-Smirnov test on the logaritmic scale is given.
}
\seealso{
\code{\link{anglePlot}}, \code{\link{PCADSC}}
}
