% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMydata.R
\name{simMyData}
\alias{simMyData}
\title{import lava
importFrom stats simulate rnorm var}
\usage{
simMyData(lavaObj = NULL, nObs = NULL, n1 = NULL, n2 = NULL,
  biasSize = NULL, affectDegree = NULL, varSize = NULL,
  nClusters = NULL)
}
\description{
import lava
importFrom stats simulate rnorm var
}

