% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallyPancake.R
\name{wallyPancake}
\alias{wallyPancake}
\title{Wally plot for a PCADSC object}
\usage{
wallyPancake(PCADSCobj, data, nrow, ncol, covCO = 1, varAnnotation = "cum",
  useComps = NULL)
}
\arguments{
\item{PCADSCobj}{A PCADSC object, as created by \code{\link{makePCADSC}}.}

\item{data}{The dataset used to generate the PCADSC object.}

\item{nrow}{The number of rows of plots in the outputted Wally plot.}

\item{ncol}{The number of columns of plo0ts in the outputted Wally plot.}

\item{covCO}{A cummulated variance cut-off limit. This parameter controls how
many components are illustrated in each of the PCADSC plots.}

\item{varAnnotation}{...}

\item{useComps}{...}
}
\description{
Creates a Wally plot for a \code{\link{PCADSC}} object,
 i.e. a grid of plots where the grouping variable for the PCADSC procedure
 has been replaced by random grouping in all but one (the top-left) plot.
 Thereby, the plot can be used to access the variability of PCADSC results.
}

