% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglePlot.R
\name{anglePlot}
\alias{anglePlot}
\title{Angle plot}
\usage{
anglePlot(x)
}
\arguments{
\item{x}{...}
}
\description{
Compares eigenvalues and eigenvectors form two datasets. Kolmogorov-Smirnov and Cramer-von Mises tests evaluated by permutation tests might be implemented later.
}
\examples{
#load iris data
data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
iris$group <- "setosa"
iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
irisPCADSC <- PCADSC(iris, "group", setdiff(names(iris), c("group", "Species")))

#make an angle plot
anglePlot(irisPCADSC)

}

