% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromaPlot.R
\name{chromaPlot}
\alias{chromaPlot}
\title{Chroma plot}
\usage{
chromaPlot(x, varLabels = NULL, covCO = NULL, splitLabels = NULL,
  varAnnotation = "cum", useComps = NULL)
}
\arguments{
\item{x}{Either a \code{PCADSC} object or a \code{chromaInfo} object, as produced
by \code{\link{PCADSC}} and \code{\link{doChroma}}.}

\item{varLabels}{A vector of character string labels for the variables used in
\code{pcadscObj}. If non-null, these labels appear in the plot instead of the
variable names. Note that they must be listed in the same order as the variables
in \code{pcadscObj} and this order can be expected by calling
\code{pcadscObj$varNames}. If \code{NULL} (the default), these variable
names are used.}

\item{covCO}{Variance contribution cut-off}

\item{splitLabels}{Labels for the two categories of the splitting variable used
to create the PCADSC object, \code{pcadscObj}, given as a named list (see examples).
These labels will appear in the headers of the two PCADSC plots.
If \code{NULL} (the default), the original levels of the splitting variable
are used.}

\item{varAnnotation}{If \code{"cum"} (the default), cummulated explained variance
percentages are annotated to the right of the bars for each component. If \code{"raw"},
the non-cummulated percentages of explained variance are added instead. If \code{NULL},
no annotation is added. Note that \code{"cum"} is only allowed if all components
are plottet (see \code{useComp}).}

\item{useComps}{A vector of integers with the numbers of the principal component
that should be included in the plot.}
}
\description{
??
}
\examples{
#load iris data
data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
iris$group <- "setosa"
iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
irisPCADSC <- PCADSC(iris, "group", setdiff(names(iris), c("group", "Species")))

#make a chroma plot
chromaPlot(irisPCADSC)


}

