% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doChroma.R
\name{doChroma}
\alias{doChroma}
\title{Compute chroma information}
\usage{
doChroma(x)
}
\arguments{
\item{x}{...}
}
\description{
Compute standardized PCA loadings and cummulative variance contributions
BLABLA
}

