% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hairplot.R
\name{hairplot}
\alias{hairplot}
\title{Hairplot}
\usage{
hairplot(data1, data2, standardize = TRUE, make.plot = TRUE,
  arrow.len = 0.05)
}
\arguments{
\item{data1}{???}

\item{data2}{???}

\item{standardize}{???}

\item{make.plot}{????}

\item{arrow.len}{???}
}
\description{
????
}
\examples{
#load iris data
  data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
 iris$group <- "setosa"
 iris$group[iris$Species != "setosa"] <- "non-setosa"

#Make hairplot
 hairplot(iris[iris$group == "setosa", 1:4], iris[iris$group == "non-setosa", 1:4])

}

