% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADSC.R, R/print.R, R/qplot.R
\docType{class}
\name{PCADSC-class}
\alias{PCADSC}
\alias{PCADSC-class}
\alias{print,PCADSC-method}
\alias{qplot,PCADSC-method}
\alias{show,PCADSC-method}
\title{A PCADSC object}
\usage{
\S4method{print}{PCADSC}(x, ...)

\S4method{show}{PCADSC}(object)

\S4method{qplot}{PCADSC}(x, varLabels = NULL, covCO = NULL,
  splitLabels = NULL)
}
\arguments{
\item{x}{A \code{PCADSC} object}

\item{...}{For internal use only}

\item{object}{A \code{PCADSC} object}

\item{varLabels}{A vector of character string labels for the variables used in
\code{pcadscObj}. If non-null, these labels appear in the plot instead of the
variable names. Note that they must be listed in the same order as the variables
in \code{pcadscObj} and this order can be expected by calling
\code{pcadscObj$varNames}. If \code{NULL} (the default), these variable
names are used.}

\item{covCO}{Not implemented, something like a variance contribution cut-off, maybe?}

\item{splitLabels}{Labels for the two categories of the splitting variable used
to create the PCADSC object, \code{pcadscObj}, given as a named list (see examples).
These labels will appear in the headers of the two PCADSC plots.
If \code{NULL} (the default), the original levels of the splitting variable
are used.}
}
\description{
Blabla
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print an overview of a PCADSC object

\item \code{show}: Show a PCADSC object. Alias for \code{print}.

\item \code{qplot}: Plot a PCADSC object
}}
\section{Slots}{

\describe{
\item{\code{pcaFrame}}{...}

\item{\code{splitBy}}{...}

\item{\code{splitLevels}}{...}

\item{\code{varNames}}{...}

\item{\code{n1}}{...}

\item{\code{n2}}{...}

\item{\code{nObs1}}{...}

\item{\code{nObs2}}{...}
}}
\examples{
#load iris data
  data(iris)

#Define grouping variable, grouping the observations by whether their species is
#Setosa or not
 iris$group <- "setosa"
 iris$group[iris$Species != "setosa"] <- "non-setosa"

#make a PCADSC object, splitting the data by "group"
 irisPCADSC <- makePCADSC(iris, "group",
                        var=setdiff(names(iris), c("group", "Species")))
#plot it
 qplot(irisPCADSC)

#print it
 print(irisPCADSC)



}

