% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCADSC.R
\name{PCADSC}
\alias{PCADSC}
\title{Compute the elements used for PCADSC}
\usage{
PCADSC(data, splitBy, var = NULL, covCO = NULL)
}
\arguments{
\item{data}{A dataset, either a \code{data.frame} or a \code{matrix} with variables
in columns and observations in rows. WHAT TO DO ABOUT DATA.TABLES?}

\item{splitBy}{A grouping variable with two levels defining the two groups within the
dataset whose data structures we wish to compare. If \code{splitBy} has more than
two levels, only the first two levels are used.}

\item{var}{The variable names in \code{data} to include in the PCADSC. If \code{NULL}
(the default), all variables except for \code{splitBy} are used.}

\item{covCO}{Not implemented yet (WHAT DID I HAVE IN MIND HERE? CUMMULATED VARIANCE
CONTRIBUTION CUT-OFF, MAYBE? SHOULDN'T BE IN THE OBJECT THOUGH, SHOULD RATHER
BE IN THE PLOTTING FUNCTION)}
}
\value{
A list including the following objects: ...
}
\description{
Calculate standardized PCA loadings and accumulated variance contributions
seperately on two parts of a dataset in order to perform non-parametric data structure
comparisons.
}
\examples{
#DO something

}

